refactor: clear warnings
