+++ src/models/message_model.rs
