+++ src/models/mod.rs
@@ -1,4 +1,4 @@
 pub mod error_response;
-pub mod messageModel;
+pub mod message_model;
 pub mod token_model;
 pub mod auth_request;
