+++ src/repositories/mod.rs
@@ -1,3 +1,3 @@
-pub mod base_Repository;
+pub mod base_repository;
 pub mod token_repository;
 pub mod credentials_repository;
