+++ src/test/static_files_test.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code, unused_imports, unused_variables)]
 use std::sync::Arc;
 use crate::config;
